loader, server-side rendering for SEO, product card styling changes
