+++ components/products-client.tsx
@@ -1,0 +1,99 @@
+"use client";
+
+import { useEffect, useMemo, useState } from "react";
+import Image from "next/image";
+import Link from "next/link";
+import { useSearchParams } from "next/navigation";
+import { Heart } from "lucide-react";
+import { Button } from "@/components/ui/button";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import { SimplifiedProduct } from "@/lib/services/api";
+import ProductCard from "@/components/product-card";
+
+interface ProductsClientProps {
+  products: SimplifiedProduct[];
+}
+
+export default function ProductsClient({ products }: ProductsClientProps) {
+  const searchParams = useSearchParams();
+  const [selectedCategory, setSelectedCategory] = useState<string>("all");
+  const [searchQuery, setSearchQuery] = useState<string>("");
+
+  useEffect(() => {
+    const category = searchParams.get("category");
+    if (category) {
+      setSelectedCategory(category.toLowerCase());
+    }
+  }, [searchParams]);
+
+  const filteredProducts = useMemo(() => {
+    return products.filter((product) => {
+      const matchesCategory =
+        selectedCategory === "all" ||
+        product.category?.toLowerCase() === selectedCategory;
+      const matchesSearch = product.name
+        .toLowerCase()
+        .includes(searchQuery.toLowerCase());
+      return matchesCategory && matchesSearch;
+    });
+  }, [products, selectedCategory, searchQuery]);
+
+  if (!products || products.length === 0) {
+    return (
+      <div className="text-center p-12">
+        <p className="text-lg mb-4">No products available.</p>
+        <Button asChild>
+          <Link href="/">Go Home</Link>
+        </Button>
+      </div>
+    );
+  }
+
+  return (
+    <>
+      <div className="mb-8 flex flex-col sm:flex-row sm:items-center sm:justify-between gap-4">
+        <input
+          type="text"
+          value={searchQuery}
+          onChange={(e) => setSearchQuery(e.target.value)}
+          placeholder="Search products by name..."
+          className="border border-gray-300 rounded-md px-4 py-2 w-full sm:w-72 focus:outline-none focus:ring-2 focus:ring-primary"
+        />
+        <Select value={selectedCategory} onValueChange={setSelectedCategory}>
+          <SelectTrigger className="w-[200px]">
+            <SelectValue placeholder="Filter by category" />
+          </SelectTrigger>
+          <SelectContent>
+            <SelectItem value="all">All Categories</SelectItem>
+            <SelectItem value="newborn">Newborn</SelectItem>
+            <SelectItem value="boy">Boy</SelectItem>
+            <SelectItem value="girl">Girl</SelectItem>
+            <SelectItem value="occasion">Occasion</SelectItem>
+            <SelectItem value="couture">Couture</SelectItem>
+          </SelectContent>
+        </Select>
+      </div>
+
+      {filteredProducts.length === 0 ? (
+        <div className="text-center p-12">
+          <p className="text-lg mb-4">No products found in this category.</p>
+          <Button asChild>
+            <Link href="/products">View all products</Link>
+          </Button>
+        </div>
+      ) : (
+        <div className="grid grid-cols-2 lg:grid-cols-4 gap-8">
+          {filteredProducts.map((product) => (
+            <ProductCard key={product.id} product={product} />
+          ))}
+        </div>
+      )}
+    </>
+  );
+}
